Merge in TAN/tan-client-experience from bugfix/VAN-27534-qa-settings-sometimes-smart-assist-page-link-can-not-show to develop

* commit 'd313ff846f86eb001c9fe25556d03b2835794a99':
  fix the router path
  VAN-27534 bug fix

+++ vtr-ui/src/app/components/pages/page-device-settings/page-device-settings.component.ts
@@ -377,25 +377,27 @@ export class PageDeviceSettingsComponent implements OnInit, OnDestroy {
 	}
 
 	getSmartAssistCapability(isAvaliable: boolean) {
+		// === from cache ===
 		let smartAssistCap = this.localCacheService.getLocalCacheValue(
 			LocalStorageKey.IsSmartAssistSupported
 		);
-		if (smartAssistCap !== undefined) {
-			if (!smartAssistCap) {
-				this.menuItems = this.commonService.removeObjById(this.menuItems, 'smart-assist');
-			}
+		if (smartAssistCap !== undefined && !smartAssistCap) {
+			this.menuItems = this.commonService.removeObjById(this.menuItems, 'smart-assist');
 		}
 
-		smartAssistCap = isAvaliable;
-		if (!smartAssistCap) {
+		// === from the real data ===
+		// filter out the invalid value
+		if (isAvaliable === null) return;
+
+		if (!isAvaliable) {
 			this.menuItems = this.commonService.removeObjById(this.menuItems, 'smart-assist');
 		} else {
 			const smartAssistMenuItem = this.menuItems.find(item => item.id === 'smart-assist');
 			if (!smartAssistMenuItem) {
 				this.menuItems.push({
 					id: 'smart-assist',
-					label: 'device.smartAssist.title',
-					path: 'device-settings/smart-assist',
+					label: this.translate.instant('device.smartAssist.title'),
+					path: '/device/device-settings/smart-assist',
 					icon: 'smart-assist',
 					iconClass: 'icomoon-Smart-Assist',
 					canDeactivate: [GuardService],
